+++ nbody/modules/core/mylinal.py
