+++ nbody/modules/core/mymath.py
